Payments/src/requirements.txt
Django==1.9
dj-database-url==0.3.0
dj-static==0.0.6
django-appconf==0.6
djangorestframework
drf-nested-routers==0.9.0
requests>=2.5.1
